admin see all bookmark

+++ app/Http/Controllers/BookmarkController.php
@@ -16,7 +16,14 @@ class BookmarkController extends Controller
     public function index()
     {
         $user = Auth::user();
-        $bookmarks = Bookmark::Where('user_id', $user->id)->orWhere('public', true)->paginate(15);
+        if($user->hasRole('Admin'))
+        {
+            $bookmarks = Bookmark::paginate(15);
+        }
+        else
+        {
+            $bookmarks = Bookmark::Where('user_id', $user->id)->orWhere('public', true)->paginate(15);
+        }
         return view('bookmark.index', compact('bookmarks'));
     }
 
@@ -77,7 +84,8 @@ class BookmarkController extends Controller
      */
     public function edit(Bookmark $bookmark)
     {
-        abort_if($bookmark->user_id !== auth()->id(), 403);
+        $user = Auth::user();
+        abort_if($bookmark->user_id !== $user->id||!$user->hasRole('Admin'), 403);
         return view('bookmark.edit',  compact('bookmark'));
     }
 
@@ -90,6 +98,8 @@ class BookmarkController extends Controller
      */
     public function update(Request $request, Bookmark $bookmark)
     {
+        $user = Auth::user();
+        abort_if($bookmark->user_id !== $user->id||!$user->hasRole('Admin'), 403);
         $request->validate([
             'name' => ['required', 'max:255'],
             'link' => ['required', 'max:512'],
@@ -116,6 +126,8 @@ class BookmarkController extends Controller
      */
     public function destroy(Bookmark $bookmark)
     {
+        $user = Auth::user();
+        abort_if($bookmark->user_id !== $user->id||!$user->hasRole('Admin'), 403);
         $bookmark->detag();
         $bookmark->delete();
         return redirect('/profile');
